+++ src/ops/IdpOps.ts
@@ -1,4 +1,4 @@
-import { SocialIdpSkeleton } from "@rockcarver/frodo-lib/types/api/ApiTypes";
+import { SocialIdpSkeleton } from '@rockcarver/frodo-lib/types/api/ApiTypes';
 
 /**
  * Get a one-line description of the social idp object
@@ -8,4 +8,29 @@ import { SocialIdpSkeleton } from "@rockcarver/frodo-lib/types/api/ApiTypes";
 export function getOneLineDescription(socialIdpObj: SocialIdpSkeleton): string {
   const description = `[${socialIdpObj._id['brightCyan']}] ${socialIdpObj._type._id}`;
   return description;
+}
+
+/**
+ * Get markdown table header
+ * @returns {string} markdown table header
+ */
+export function getTableHeaderMd(): string {
+  let markdown = '';
+  markdown += '| Name/Id | Status | Type |\n';
+  markdown += '| ------- | ------ | ---- |';
+  return markdown;
+}
+
+/**
+ * Get a table-row of the social idp in markdown
+ * @param {SocialIdpSkeleton} socialIdpObj social idp object to describe
+ * @returns {string} a table-row of the social idp in markdown
+ */
+export function getTableRowMd(socialIdpObj: SocialIdpSkeleton): string {
+  const row = `| ${socialIdpObj._id} | ${
+    socialIdpObj.enabled === false
+      ? ':o: `disabled`'
+      : ':white_check_mark: `enabled`'
+  } | ${socialIdpObj._type.name} |`;
+  return row;
 }