+++ src/ops/Saml2Ops.ts
@@ -22,3 +22,30 @@ export function getOneLineDescription(
   }${roles.length ? ' ' + roles.join(', ') + ')' : ')'}`;
   return description;
 }
+
+/**
+ * Get markdown table header
+ * @returns {string} markdown table header
+ */
+export function getTableHeaderMd(): string {
+  let markdown = '';
+  markdown += '| Entity Id | Location | Role(s) |\n';
+  markdown += '| --------- | -------- | ------- |';
+  return markdown;
+}
+
+/**
+ * Get a table-row of the saml2 provider in markdown
+ * @param {Saml2ProviderSkeleton} saml2ProviderObj saml2 provider object to describe
+ * @returns {string} a table-row of the saml2 provider in markdown
+ */
+export function getTableRowMd(saml2ProviderObj: Saml2ProviderSkeleton): string {
+  const roles: string[] = [];
+  for (const [key, value] of Object.entries(roleMap)) {
+    if (saml2ProviderObj[key]) {
+      roles.push(value);
+    }
+  }
+  const row = `| ${saml2ProviderObj.entityId} | ${saml2ProviderObj.entityLocation} | ${roles.length ? roles.join(', ') : ''} |`;
+  return row;
+}